new page for mars project
redisigning project so that i can showcase better design skills as well as using bootstrap in a project. Got basic html down and centered things. The arrows prev and next are causing a little problem so need to fix that next time, want them positioned on left and right of screen.